requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
slowapi==0.1.9
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
aiohttp==3.9.1

# LangGraph Multi-Agent System
langgraph==0.2.28
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.6

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.11.4
pytesseract==0.3.13

# HTML Processing
beautifulsoup4==4.14.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# AI/ML
tiktoken==0.8.0
